refactor: splitting bi tests into csv and excel files
